Shopping_Agent_Using_Gemini_AI/pyproject.toml
[project]
name = "shopping-agent-using-gemini-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.1",
    "python-dotenv>=1.1.1",
]
